third_party/WebKit/LayoutTests/crypto/generateKey-expected.txt

CONSOLE ERROR: Invalid keyUsages argument
CONSOLE ERROR: Algorithm: AES-CBC: AesKeyGenParams: length: Missing required property
CONSOLE ERROR: Algorithm: AES-CBC: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CBC: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CBC: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CTR: AesKeyGenParams: length: Missing required property
CONSOLE ERROR: Algorithm: AES-CTR: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CTR: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CTR: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-KW: AesKeyGenParams: length: Missing required property
CONSOLE ERROR: Algorithm: AES-KW: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-KW: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-KW: AesKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: HMAC: HmacKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: HMAC: HmacKeyGenParams: hash: Algorithm: Unrecognized algorithm name
CONSOLE ERROR: Algorithm: HMAC: HmacKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: HMAC: HmacKeyGenParams: length: Is not a number
CONSOLE ERROR: Algorithm: HMAC: HmacKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: HMAC: HmacKeyGenParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: HMAC: HmacKeyGenParams: length: Is not a number
CONSOLE ERROR: Algorithm: HMAC: HmacKeyGenParams: length: Is not a number
CONSOLE ERROR: Invalid key length: it is either zero or not a multiple of 8 bits
CONSOLE ERROR: Algorithm: RSASSA-PKCS1-v1_5: RsaHashedKeyGenParams: modulusLength: Outside of numeric range
CONSOLE ERROR: Algorithm: RSASSA-PKCS1-v1_5: RsaHashedKeyGenParams: modulusLength: Is not a number
CONSOLE ERROR: Algorithm: RSASSA-PKCS1-v1_5: RsaHashedKeyGenParams: modulusLength: Missing required property
CONSOLE ERROR: Algorithm: RSASSA-PKCS1-v1_5: RsaHashedKeyGenParams: publicExponent: Missing or not a Uint8Array
CONSOLE ERROR: Algorithm: RSASSA-PKCS1-v1_5: RsaHashedKeyGenParams: publicExponent: Missing or not a Uint8Array
CONSOLE ERROR: Algorithm: RSASSA-PKCS1-v1_5: RsaHashedKeyGenParams: publicExponent: Missing or not a Uint8Array
CONSOLE ERROR: Algorithm: RSAES-PKCS1-v1_5: RsaKeyGenParams: modulusLength: Outside of numeric range
CONSOLE ERROR: Algorithm: RSAES-PKCS1-v1_5: RsaKeyGenParams: modulusLength: Is not a number
CONSOLE ERROR: Algorithm: RSAES-PKCS1-v1_5: RsaKeyGenParams: modulusLength: Missing required property
CONSOLE ERROR: Algorithm: RSAES-PKCS1-v1_5: RsaKeyGenParams: publicExponent: Missing or not a Uint8Array
CONSOLE ERROR: Algorithm: RSAES-PKCS1-v1_5: RsaKeyGenParams: publicExponent: Missing or not a Uint8Array
CONSOLE ERROR: Algorithm: RSAES-PKCS1-v1_5: RsaKeyGenParams: publicExponent: Missing or not a Uint8Array
Tests crypto.subtle.generateKey.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.generateKey(aesCbc, extractable, -1) threw exception TypeError: The 3rd argument is neither an array, nor does it have indexed properties..
PASS crypto.subtle.generateKey(aesCbc, extractable, null) threw exception TypeError: The 3rd argument is neither an array, nor does it have indexed properties..
PASS: 'crypto.subtle.generateKey(aesCbc, extractable, ['boo'])' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-cbc' }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-cbc', length: 70000 }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-cbc', length: -3 }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-cbc', length: -Infinity }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-ctr' }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-ctr', length: 70000 }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-ctr', length: -3 }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-ctr', length: -Infinity }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-kw' }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-kw', length: 70000 }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-kw', length: -3 }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({ name: 'aes-kw', length: -Infinity }, extractable, keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: -3}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: ''}, length: 48}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: 5000000000}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: NaN}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: Infinity}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: -Infinity}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: crypto}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: undefined}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: null}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', modulusLength: -30}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', modulusLength: NaN}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5'}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', modulusLength: 10}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', modulusLength: 10, publicExponent: 10}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', modulusLength: 10, publicExponent: null}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5', modulusLength: -30}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5', modulusLength: NaN}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5'}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5', modulusLength: 10}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5', modulusLength: 10, publicExponent: 10}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'RSAES-PKCS1-v1_5', modulusLength: 10, publicExponent: null}, extractable , keyUsages)' rejected with null
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: '48'}, extractable , keyUsages)' accepted with [object Key]
PASS: 'crypto.subtle.generateKey({name: 'hmac', hash: {name: 'sha-256'}, length: true}, extractable , keyUsages)' accepted with [object Key]
PASS successfullyParsed is true

TEST COMPLETE
